% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ssofr}
\alias{plot.ssofr}
\title{Plot output from ssofr object}
\usage{
\method{plot}{ssofr}(x, levels = c(0.025, 0.975), which = 1:2, ids = 1:3)
}
\arguments{
\item{x}{\code{ssofr} object}

\item{levels}{lower and upper quantiles for the credible intervals}

\item{which}{subset of \code{1:2} specifying which outputs to plot. \code{1} shows the regression coefficient function and \code{2} shows the fitted functional covariate curves.}

\item{ids}{integer vector specifying index of which fitted functional covariate curves to plot.}
}
\value{
Plots of the posterior mean and pointwise credible interval for the regression coefficient function, as well as the fitted curves if \code{fittedx} is specified.
}
\description{
Plots summary of SOFR model or sparse curve fitting model using the MCMC draws from \code{ssofr} or \code{sffm}.
}
