% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sffm.R
\name{sffm}
\alias{sffm}
\title{Sparse functional factor model}
\usage{
sffm(Ym, K = 10, S = 2000, S_burn = S/2, sparse = TRUE)
}
\arguments{
\item{Ym}{matrix of functional responses. Each column corresponds to one function and each row is a measurement at a single time point on a common grid. May contain NAs.}

\item{K}{number of basis functions.}

\item{S}{number of total MCMC iterations}

\item{S_burn}{number of initial MCMC iterations to discard for warm-up}

\item{sparse}{set to TRUE if functions considered sparsely observed.}
}
\value{
A list with the following elements:
\itemize{
\item X: the inputted design matrix
}
}
\description{
Performs MCMC estimation for Bayesian functional factor model.
}
