% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sf_data.R
\name{sim_sf_data}
\alias{sim_sf_data}
\title{Simulated Functional Data}
\usage{
sim_sf_data(N, Tn, K = 4, sig_noise = 1e-07, sig_alpha = 1, seed = NULL)
}
\arguments{
\item{N}{number of subjects.}

\item{Tn}{total number of time points.}

\item{K}{number of basis functions}

\item{sig_noise}{observation level variance.}

\item{sig_alpha}{basis coefficients variance.}

\item{seed}{(optional) sets the seed.}
}
\value{
list with the following elements:
\itemize{
\item Y: simulated functional data with noise
\item Y_true: simulated true functional data
\item B: true basis used to generate functions
}
}
\description{
Generate simulated functional data.
}
