% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssofr.R
\name{ssofr}
\alias{ssofr}
\title{Sparse scalar-on-function regression}
\usage{
ssofr(
  y,
  x,
  w = NULL,
  Tau,
  K = 10,
  Kphi = 15,
  S = 2000,
  S_burn = 1000,
  sparse = TRUE,
  showprogress = TRUE
)
}
\arguments{
\item{y}{vector of scalar responses.}

\item{x}{matrix of functional responses. Each column corresponds to one function and each row is a measurement at a single time point on a common grid. May contain NAs.}

\item{w}{design matrix of scalar covariates using \code{model.matrix}. It should include column of 1s for the intercept.}

\item{Tau}{vector specifying common grid of time points on which function is observed.}

\item{K}{number of basis functions in factor model.}

\item{Kphi}{number of basis functions for functional coefficient.}

\item{S}{number of total MCMC iterations.}

\item{S_burn}{number of initial MCMC iterations to discard for warm-up.}

\item{sparse}{set to TRUE if functions considered sparsely observed.}

\item{showprogress}{show progress of MCMC algorithm. Defaults to TRUE.}
}
\value{
A list with the following elements:
\itemize{
\item X: the inputted design matrix
}
}
\description{
Performs MCMC estimation for Bayesian scalar-on-function regression with sparse, noisy functional covariate.
}
